Renamed MarkerDetailsController to one with suffix old
